Handle images as floats rather than int16_t
Also, remove bloom - it's not a useful model for what really happens and takes
too long (this isn't a detector simulation..)

+++ src/image.h
@@ -70,7 +70,7 @@ struct rvec
 struct image {
 
 	int			*hdr;      /* Actual counts */
-	int16_t			*data;     /* Integer counts after bloom */
+	float			*data;     /* Integer counts after bloom */
 	double complex		*sfacs;
 	double			*twotheta;
 	struct molecule		*molecule;
